LayoutTests/inspector/timeline/timeline-paint-expected.txt

Tests the Timeline API instrumentation of a paint event


Paint Properties:
{
    children : <object>
    counters : <object>
    data : {
        height : <number>
        width : <number>
        x : 0
        y : 0
    }
    endTime : <number>
    frameId : <string>
    stackTrace : <object>
    startTime : <number>
    type : "Paint"
    usedHeapSize : <number>
}
